core: Materialize Toast notifications
Toasts were never materialized by Google as they discouraged their
use, promoting SnackBars instead. Despite this, many developers
(including Google) continue to use toasts.

This commit replaces the standard toast with a material card
containing the app icon. Icon code was taken from 0xD34D's commit
and modified to be unpadded and centered vertically to be inline
with the text. It also replaces the fade animations with slide
animations to be more material.

The toast card color is themeable via the "toast_bg" color value,
and the text color is themeable via "toast_text". The shadow of
the card is themeable via "toast_elevation"

Change-Id: I47577ecb31c2f292cad3ca3e47e94f39edc6cbef
Signed-off-by: Joe Maples <joe@frap129.org>
Signed-off-by: Genkzsz11 <Genkzsz11@gmail.com>
Signed-off-by: Reza Adi Pangestu <rezaadipangestu5@gmail.com>
Signed-off-by: HeroBuxx <herobuxx@gmail.com>